- Spawning: Integrate bits of FDA's Spawn Handler
Technically a "Step 1 of 3" for this. Next is the menu and keyconf, then final is adding last bits and making sure everything works.

This now makes it so all of the separate weapons and the beacon itself can, optionally, spawn also. Radical.

Also makes it so the weapons actually make it so you don't drop their respective mag whilst holding them and you purge ammo. Cool.